+++ src/main/scala/org/podval/docbook/gradle/DocBook2.scala
@@ -5,11 +5,12 @@ import org.gradle.api.file.CopySpec
 import java.io.File
 import javax.xml.transform.stream.{StreamResult, StreamSource}
 import org.apache.tools.ant.filters.ReplaceTokens
-import org.gradle.api.artifacts.Configuration
 import org.xml.sax.InputSource
 import scala.collection.JavaConverters._
 
 abstract class DocBook2 {
+
+  final def name: String = Util.className(this).toUpperCase
 
   final def run(
     layout: Layout,
@@ -20,21 +21,32 @@ abstract class DocBook2 {
     project: Project,
     logger: Logger
   ): Unit = {
-    logger.info(s"\nProcessing DocBook to $finalOutputFormat")
-    val saxonOutputDirectory: File = intermediateOutputDirectory(layout)
+    logger.info(s"\nProcessing DocBook to $name")
+
+    val saxonOutputDirectory: File = new File(
+      if (saxon2intermediate) layout.intermediateOutputDirectoryRoot else layout.finalOutputDirectoryRoot,
+      saxonOutputFormat
+    )
     saxonOutputDirectory.mkdirs
 
-    val saxonOutputFile: File = intermediateOutputFile(layout, inputFileName)
+    val saxonOutputFile: File = outputFileFor(saxonOutputDirectory, saxonOutputFormat, inputFileName)
 
-    val allAdditionalParameters: Map[String, String] =
+    val additionalParameters: Map[String, String] =
       Map("img.src.path" -> (layout.imagesDirectoryName + "/")) ++
-        additionalParameters(layout, inputFileName)
+        (if (!usesHtml) Map.empty else Map(
+          "base.dir" -> saxonOutputDirectory.getAbsolutePath,
+          "root.filename" -> Util.fileNameWithoutExtension(saxonOutputFile),
+          "html.stylesheet" -> (layout.cssDirectoryName + "/" + layout.cssFileName)
+        ))
+
+    val xslParametersEffective: Map[String, String] =
+      xslParameters ++ (additionalParameters -- xslParameters.keySet)
 
     saxon.run(
       inputSource = new InputSource(layout.inputFile(inputFileName).toURI.toASCIIString),
       stylesheetSource = new StreamSource(layout.stylesheetFile(saxonOutputFormat)),
       outputTarget = new StreamResult(saxonOutputFile),
-      xslParameters = xslParameters ++ (allAdditionalParameters -- xslParameters.keySet)
+      xslParameters = xslParametersEffective
     )
 
     logger.info(s"Copying images")
@@ -47,17 +59,41 @@ abstract class DocBook2 {
       }
     })
 
-    val outputDirectory: File = finalOutputDirectory(layout)
-    outputDirectory.mkdirs
+    if (usesHtml) {
+      val substitutionsEffective: Map[String, String] = substitutions ++ xslParametersEffective
+      logger.info(s"Copying CSS")
+      project.copy(new Action[CopySpec] {
+        override def execute(copySpec: CopySpec): Unit = {
+          copySpec
+            .into(saxonOutputDirectory)
+            .from(layout.cssDirectory.getParentFile)
+            .include(layout.cssDirectoryName + "/**")
+            .filter(Map("tokens" -> substitutionsEffective.asJava).asJava, classOf[ReplaceTokens])
+        }
+      })
+    }
 
-    postProcess(
-      layout = layout,
-      inputFileName = inputFileName,
-      substitutions = substitutions ++ xslParameters,
-      project = project,
-      logger = logger
-    )
+    if (saxon2intermediate) {
+      logger.info(s"Post-processing $name")
+      val outputDirectory: File = new File(layout.finalOutputDirectoryRoot, finalOutputFormat)
+      outputDirectory.mkdirs
+
+      postProcess(
+        layout = layout,
+        saxonOutputDirectory = saxonOutputDirectory,
+        saxonOutputFile = saxonOutputFile,
+        outputFile = outputFileFor(outputDirectory, finalOutputFormat, inputFileName),
+        logger = logger
+      )
+    }
   }
+
+  private def outputFileFor(outputDirectory: File, outputFormat: String, inputFileName: String): File = {
+    val outputFileName: String = outputFileNameOverride.getOrElse(inputFileName)
+    new File(outputDirectory, outputFileName + "." + outputFormat)
+  }
+
+  def usesHtml: Boolean
 
   def saxon2intermediate: Boolean
 
@@ -65,121 +101,75 @@ abstract class DocBook2 {
 
   def saxonOutputFormat: String = finalOutputFormat
 
-  protected final def intermediateOutputDirectory(layout: Layout): File = new File(
-    if (saxon2intermediate) layout.intermediateOutputDirectoryRoot else layout.finalOutputDirectoryRoot,
-    saxonOutputFormat
-  )
-
-  protected final def finalOutputDirectory(layout: Layout): File =
-    new File(layout.finalOutputDirectoryRoot, finalOutputFormat)
-
-  protected final def intermediateOutputFile(layout: Layout, inputFileName: String): File =
-    outputFile(intermediateOutputDirectory(layout), saxonOutputFormat, inputFileName)
-
-  protected final def finalOutputFile(layout: Layout, inputFileName: String): File =
-    outputFile(finalOutputDirectory(layout), finalOutputFormat, inputFileName)
-
-  private def outputFile(outputDirectory: File, outputFormat: String, inputFileName: String): File =
-    new File(outputDirectory, outputFileName(inputFileName) + "." + outputFormat)
-
-  protected final def outputFileName(inputFileName: String): String = outputFileNameOverride.getOrElse(inputFileName)
-
   protected def outputFileNameOverride: Option[String] = None
-
-  protected def additionalParameters(layout: Layout, inputFileName: String): Map[String, String] = Map.empty
-
-  protected final def additionalParametersHtml(layout: Layout, inputFileName: String): Map[String, String] = Map(
-    "base.dir" -> intermediateOutputDirectory(layout).getAbsolutePath,
-    "root.filename" -> Util.fileNameWithoutExtension(intermediateOutputFile(layout, inputFileName)),
-    "html.stylesheet" -> (layout.cssDirectoryName + "/" + layout.cssFileName)
-  )
 
   protected def postProcess(
     layout: Layout,
-    inputFileName: String,
-    substitutions: Map[String, String],
-    project: Project,
+    saxonOutputDirectory: File,
+    saxonOutputFile: File,
+    outputFile: File,
     logger: Logger
   ): Unit = {
-  }
-
-  protected final def copyCss(
-    layout: Layout,
-    directory: File,
-    substitutions: Map[String, String],
-    project: Project,
-    logger: Logger
-  ): Unit = {
-    logger.info(s"Copying CSS")
-    project.copy(new Action[CopySpec] {
-      override def execute(copySpec: CopySpec): Unit = {
-        copySpec
-          .into(directory)
-          .from(layout.cssDirectory.getParentFile)
-          .include(layout.cssDirectoryName + "/**")
-          .filter(Map("tokens" -> substitutions.asJava).asJava, classOf[ReplaceTokens])
-      }
-    })
   }
 }
 
 object DocBook2 {
 
-  val processors: List[DocBook2] = List(DocBook2Html, DocBook2Epub, DocBook2Pdf)
-
-  def process(
-    outputFormats: List[String],
-    layout: Layout,
-    inputFileName: String,
-    xslParameters: Map[String, String],
-    substitutions: Map[String, String],
-    project: Project,
-    logger: Logger
-  ): Unit = {
-    val processorsToRun: List[DocBook2] = outputFormats.map { outputFormat =>
-      val processor: Option[DocBook2] = processors.find(_.finalOutputFormat == outputFormat)
-      if (processor.isEmpty) {
-        val message: String =
-          s"""Unsupported output format $outputFormat;
-             |  supported formats are: ${processors.map(_.finalOutputFormat)}""".stripMargin
-        logger.error(message)
-        throw new IllegalArgumentException(message)
-      } else {
-        processor.get
-      }
-    }
-
-    logger.info(s"Output formats selected: ${processorsToRun.map(_.finalOutputFormat)}")
-
-    val docBookXslConfiguration: Configuration = project.getConfigurations.findByName("docBookXsl")
-
-    // Unpack DocBook XSLT stylesheets
-    if (!layout.docBookXslDirectory.exists) {
-      logger.info(s"Preparing DocBook XSLT stylesheets")
-      project.copy(new Action[CopySpec] {
-        override def execute(copySpec: CopySpec): Unit = {
-          copySpec
-            .into(layout.explodeDocBookXslInto)
-            .from(project.zipTree(docBookXslConfiguration.getSingleFile))
-        }
-      })
-    }
-
-    val saxon: Saxon = new Saxon(
-      substitutions = substitutions,
-      xslDirectory = layout.docBookXslDirectory,
-      dataDirectory = layout.dataDirectory,
-      logger: Logger
-    )
-
-    processorsToRun.foreach(_.run(
-      layout = layout,
-      inputFileName = inputFileName,
-      saxon = saxon,
-      xslParameters = xslParameters,
-      substitutions = substitutions,
-      project = project,
-      logger = logger
-    ))
+  object Html extends DocBook2 {
+    override def usesHtml: Boolean = true
+    override def saxon2intermediate: Boolean = false
+    override def finalOutputFormat: String = "html"
+    override protected def outputFileNameOverride: Option[String] = Some("index")
   }
+
+  object Pdf extends DocBook2 {
+    override def usesHtml: Boolean = false
+    override def saxon2intermediate: Boolean = true
+    override def saxonOutputFormat: String = "fo"
+    override def finalOutputFormat: String = "pdf"
+
+    override protected def postProcess(
+      layout: Layout,
+      saxonOutputDirectory: File,
+      saxonOutputFile: File,
+      outputFile: File,
+      logger: Logger
+    ): Unit = Fop.run(
+      configurationFile = layout.fopConfigurationFile,
+      inputFile = saxonOutputFile,
+      baseDirectory = saxonOutputDirectory,
+      outputFile = outputFile,
+      logger = logger
+    )
+  }
+
+  object Epub2 extends DocBook2 {
+    override def usesHtml: Boolean = true
+    override def saxon2intermediate: Boolean = true
+    override def finalOutputFormat: String = "epub"
+
+    override protected def postProcess(
+      layout: Layout,
+      saxonOutputDirectory: File,
+      saxonOutputFile: File,
+      outputFile: File,
+      logger: Logger
+    ): Unit = Epub.pack(
+      inputDirectory = saxonOutputDirectory,
+      outputFile = outputFile
+    )
+  }
+
+  private val all: List[DocBook2] = List(Html, Epub2, Pdf)
+
+  def forName(name: String): DocBook2 = all.find(_.name.toUpperCase == name.toUpperCase).getOrElse {
+    throw new IllegalArgumentException(
+      s"""Unsupported output format $name;
+         |  supported formats are: $availableFormatNames""".stripMargin
+    )
+  }
+
+  val availableFormats: Seq[String] = all.map(_.name)
+
+  val availableFormatNames: String = availableFormats.mkString(", ")
 }