+++ src/main/scala/org/podval/docbook/gradle/Util.scala
@@ -14,4 +14,8 @@ object Util {
 
   def drop(what: String, from: String): Option[String] =
     if (from.startsWith(what)) Some(from.drop(what.length)) else None
+
+  // Maybe in JDK 9 and later I won't need to deal with '$'?
+  // see https://bugs.java.com/bugdatabase/view_bug.do?bug_id=JDK-8057919
+  def className(obj: AnyRef): String = obj.getClass.getSimpleName.replace("$", "")
 }